The MONGO_INITDB_DATABASE variable is not super intuitive
This is the text from the https://hub.docker.com/_/mongo page:

MONGO_INITDB_DATABASE
This variable allows you to specify the name of a database to be used for creation scripts in /docker-entrypoint-initdb.d/*.js (see Initializing a fresh instance below). MongoDB is fundamentally designed for "create on first use", so if you do not insert data with your JavaScript files, then no database is created.

I don't think that it is needed here, so I made all of the variables contingent on there being a username provided for authentication